chore: extract `Request` and `Response` into its own module
